Fix/191 aggregate inversion solution as predecessor (#192)
* test reproduces the error
* AggregateInversionSolution added to PredecessorUnion with test cover; cast random args to int;
