+++ graphql_api/tests/hazard/test_aggregate_inversion_solution.py
@@ -161,61 +161,4 @@ class TestAggregateInversionSolution(unittest.TestCase, SetupHelpersMixin):
         self.assertEqual( node['predecessors'][0]['relationship'], 'Parent')
 
 
-    def create_aggregate_solution_with_predecessors(self, source_solutions, task_id, common_rupture_set):
-        """test helper"""
-        query = '''
-           mutation ($source_solutions: [ID]!, $produced_by: ID!, $digest: String!,
-                $file_name: String!, $file_size: BigInt!, $created: DateTime!,
-                $predecessors: [PredecessorInput], $common_rupture_set: ID!) {
-              create_aggregate_inversion_solution(
-                  input: {
-                      source_solutions: $source_solutions
-                      aggregation_fn: MEAN
-                      md5_digest: $digest
-                      file_name: $file_name
-                      file_size: $file_size
-                      created: $created
-                      produced_by: $produced_by
-                      predecessors: $predecessors
-                      common_rupture_set: $common_rupture_set
-                    })
-              {
-                ok
-                solution { id, file_name, file_size, md5_digest, post_url
-                    source_solutions { ... on Node{id} }
-                    produced_by { ... on Node{id} }
-                    predecessors {
-                        id,
-                        typename,
-                        depth,
-                        relationship
-                        node {
-                            ... on FileInterface {
-                                meta {k v}
-                                file_name
-                            }
-                        }
-                    }
-                }
-              }
-            }'''
-
-        # from hashlib import sha256, md5
-        filedata = BytesIO("a line\nor two".encode())
-        digest = "sha256(filedata.read()).hexdigest()"
-        filedata.seek(0) #important!
-        size = len(filedata.read())
-        filedata.seek(0) #important!
-
-        predecessors = [dict(id=source_solutions[0], depth=-1)]
-
-        variables = dict(source_solutions=source_solutions, produced_by=task_id,
-            file=filedata, digest=digest, file_name="alineortwo.txt", file_size=size,
-            predecessors=predecessors, common_rupture_set=common_rupture_set)
-        variables['created'] = dt.datetime.now(tzutc()).isoformat()
-        result = self.client.execute(query, variable_values=variables )
-        print(result)
-        return result
-
-
